requirements.txt
SpeechRecognition
pyttsx3
google-generativeai
audio_recorder_streamlit
streamlit_float
langchain
dotenv